#11 - Move atoms behind compression slider, make atom positions relative to circle centre

+++ Source/Level3Editor.h
@@ -36,7 +36,7 @@ public:
 
     void paint(Graphics&) override;
     void resized() override;
-    void resize(Rectangle<int>);
+    void resize(Rectangle<float>);
 
     void changeListenerCallback(ChangeBroadcaster* source) override;
